Refactor FilmCategoryPK setter to accept int and cast to byte; add @Service annotation to ActorServiceImpl; implement unit tests for FilmText, FilmActor, Category, Actor, and FilmActorPK entities

+++ catalogo/src/main/java/cap/maria/catalogo/Entities/FilmCategoryPK.java
@@ -29,8 +29,8 @@ public class FilmCategoryPK implements Serializable {
 	public byte getCategoryId() {
 		return this.categoryId;
 	}
-	public void setCategoryId(byte categoryId) {
-		this.categoryId = categoryId;
+	public void setCategoryId(int categoryId) {
+		this.categoryId = (byte) categoryId;
 	}
 
 	public boolean equals(Object other) {
